Table sorting: be explicit with column types
Previously, the tablesorter was detecting numeric values with thousands
separators as "text" and sorting them incorrectly.

I added thousands separators for bytes values, so they are easier to
read at a glance as the numbers get long but the sorter hadn't been
parsing them well.

Note: I'm not sure why the "Cost" column was set to "sorter: false" on
the subtables, it had still been sorting but it had been falling back to
auto-detecting the sort type.